Fix local copy of FreeRTOS-Kernel to V11.0.1
By default the build will use the local dependencies.
This will simplify things When the time comes to update, or if the need arises to use a different version of FreeRTOS-Kernel than 11.0.1.
The integrator has the option of updating the version of the FreeRTOS-Kernel submodule, or updating the dependencies.cmake file to go grab the desired version from Github and update the CMakePresets.json to have the `default` preset inherit from base rather than local_deps.